Move soil component to FF16_Environment
