+++ NAMESPACE
@@ -54,7 +54,6 @@ export(K93_test_environment)
 export(No_Disturbance)
 export(OdeControl)
 export(Water_Cohort)
-export(Water_Environment)
 export(Water_Individual)
 export(Water_Parameters)
 export(Water_Patch)
