+++ src/plant_tools.cpp
@@ -25,8 +25,6 @@ plant::Internals K93_oderunner_plant_internals(
 
 // [[Rcpp::export]]
 plant::Internals Water_oderunner_plant_internals(
-  const plant::ode::Runner<plant::tools::IndividualRunner<plant::Water_Strategy, plant::Water_Environment>>& obj) {
+  const plant::ode::Runner<plant::tools::IndividualRunner<plant::Water_Strategy, plant::FF16_Environment>>& obj) {
   return obj.obj.plant.r_internals();
 }
-
-
