+++ src/water_cohort.cpp
@@ -4,11 +4,11 @@
 // Helpers for Water model
 
 // [[Rcpp::export]]
-plant::CohortSchedule cohort_schedule_default__Parameters___Water__Water_Env(const plant::Parameters<plant::Water_Strategy,plant::Water_Environment>& p) {
-   return plant::cohort_schedule_default<plant::Parameters<plant::Water_Strategy,plant::Water_Environment> >(p);
+plant::CohortSchedule cohort_schedule_default__Parameters___Water__FF16_Env(const plant::Parameters<plant::Water_Strategy,plant::FF16_Environment>& p) {
+   return plant::cohort_schedule_default<plant::Parameters<plant::Water_Strategy,plant::FF16_Environment> >(p);
 }
 
 // [[Rcpp::export]]
-plant::CohortSchedule make_cohort_schedule__Parameters___Water__Water_Env(const plant::Parameters<plant::Water_Strategy,plant::Water_Environment>& p) {
-   return plant::make_cohort_schedule<plant::Parameters<plant::Water_Strategy,plant::Water_Environment> >(p);
+plant::CohortSchedule make_cohort_schedule__Parameters___Water__FF16_Env(const plant::Parameters<plant::Water_Strategy,plant::FF16_Environment>& p) {
+   return plant::make_cohort_schedule<plant::Parameters<plant::Water_Strategy,plant::FF16_Environment> >(p);
 }
