+++ tests/testthat/test-control.R
@@ -6,10 +6,12 @@ test_that("Defaults", {
     cohort_gradient_direction = 1L,
     cohort_gradient_richardson = FALSE,
     cohort_gradient_richardson_depth = 4, # size_t, so not int
+    
     environment_light_max_depth= 16, # size_t
     environment_light_nbase = 17, # size_t
     environment_light_tol = 1e-6,
     environment_rescale_usually = FALSE,
+    
     ode_a_dydt = 0.0,
     ode_a_y = 1.0,
     ode_step_size_initial = 1e-6,
@@ -17,11 +19,13 @@ test_that("Defaults", {
     ode_step_size_min = 1e-6,
     ode_tol_abs = 1e-6,
     ode_tol_rel = 1e-6,
+    
     plant_assimilation_adaptive = TRUE,
     plant_assimilation_iterations = 1000, # size_t so not int
     plant_assimilation_rule = 21, # size_t so not int
     plant_assimilation_over_distribution = FALSE,
     plant_assimilation_tol = 1e-6,
+    
     plant_seed_iterations = 1000, # size_t
     plant_seed_tol = 1e-8, # 1e-6, Had to change this...
